# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Documentation",
        "description": "Admin OpenApi description",
        "contact": {
            "email": "danh.le1997@hcmut.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Admin API server"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of events",
                "description": "Returns list of events",
                "operationId": "getEventsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Add a new event",
                "operationId": "addEvent",
                "requestBody": {
                    "description": "Event object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Show Event Detail",
                "description": "Returns Event",
                "operationId": "getEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update Event",
                "description": "Returns use data",
                "operationId": "updateEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Event object that needs to be added to the update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete event",
                "description": "Returns use data",
                "operationId": "deleteEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/events/{id}/tasks": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Show Tasks Belong To Event ID",
                "description": "Returns task list data",
                "operationId": "event_tasks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "customer_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "time_start": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "time_end": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "result": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "name": "api_key",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ]
}